+++ zgmtls/gm_handshake_server.go
@@ -440,7 +440,6 @@ func (hs *serverHandshakeStateGM) doFullHandshake() error {
 
 	msg, err := c.readHandshake()
 	if err != nil {
-		fmt.Println("readHandshake error:", err)
 		return err
 	}
 
@@ -562,9 +561,8 @@ func (hs *serverHandshakeStateGM) establishKeys() error {
 func (hs *serverHandshakeStateGM) readFinished(out []byte) error {
 	c := hs.c
 
-	c.readRecord(recordTypeChangeCipherSpec)
-	if c.in.err != nil {
-		return c.in.err
+	if err := c.readRecord(recordTypeChangeCipherSpec); err != nil {
+		return err
 	}
 
 	if hs.hello.nextProtoNeg {